unit tests improvements
