+++ examples/index.html
@@ -4,8 +4,8 @@
     <meta charset="utf-8">
     <title>Adobe Data Layer | Examples</title>
     <script>adobeDataLayer = [];</script>
-    <script src="js/datalayer.mocks.2.js"></script>
     <script src="../dist/adobe-client-data-layer.min.js" async></script>
+    <script src="js/datalayer.mocks.1.js"></script>
 </head>
 <body>
     <h1>Adobe Client Data Layer | Examples</h1>
